Updated SwitchType to "Blinds Percentage Inverted"
Updated SwitchType and logic to make the Slide plugin compatible with Homebridge where curtain states were inverted. E.g. closed when open and vice versa.

Replace plugin.py and update the slide switches in Domoticz to Blinds Percertage Inverted for existing slides. Be aware of use in scripts as the working is inverted ;)